requirements.txt
flask==2.0.1
pyyaml==6.0
numpy==1.21.6
opencv-python-headless==4.5.5.64
face_recognition==1.3.0
pandas==1.3.5
reportlab==3.6.8
gunicorn==20.1.0
pytest==7.1.2
